Be consistent in use of single-precision floats - MLLR-SAT (if anyone was using it) would have been broken due to this.
git-svn-id: svn+ssh://svn.code.sf.net/p/cmusphinx/code/trunk/SphinxTrain@5826 94700074-3cef-4d97-a70e-9c8c206c02f5

+++ include/s3/mllr.h
@@ -53,8 +53,8 @@ int32 regmat_read (
 
 
 int32 compute_mllr (
-             float64 *****regl,
-             float64 ****regr,
+             float32 *****regl,
+             float32 ****regr,
              const uint32  *veclen,
              uint32  nclass,
              uint32  nfeat,
@@ -65,9 +65,9 @@ int32 compute_mllr (
             );
 
 int32 solve  (
-        float64 **regl,
-        float64 *regr,
-        float64 *x,
+        float32 **regl,
+        float32 *regr,
+        float32 *x,
         int32   len
        );
